// src/lib/types.ts
// export type Digit =
//   | "0"
//   | "1"
//   | "2"
//   | "3"
//   | "4"
//   | "5"
//   | "6"
//   | "7"
//   | "8"
//   | "9"
//   | ".";
export type Digit = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9;
export type Operator =
  | "."
  | "+"
  | "-"
  | "×"
  | "÷"
  | ""
  | "AC"
  | "%"
  | "+/-"
  | "/"
  | "*"
  | "=";